requirement.txt
Django==1.7
django-appconf==1.0.2
django-ckeditor==5.0.3
django-cleanup==0.4.2
django-grappelli==2.7.3
django-imagekit==4.0.2
django-solo==1.1.5
gunicorn==19.10.0
pilkit==2.0
Pillow==6.2.2
six==1.16.0
Unidecode==1.2.0
whitenoise==3.0
